new edits to hw2
